Fix: api-data 타입 정의를 수정한다.
