+++ src/types/api-data/portfolio.ts
@@ -1,3 +1,5 @@
+import { Commission } from "@/types/api-data/commission";
+
 export type Portfolios = {
 	[key in string]: Portfolio;
 };
@@ -14,7 +16,7 @@ export type Portfolio = {
 	content: string,
 	summary: string,
 	createdAt: string,
-	section: string,
+	section: Section,
 	category: string,
 	tag: string[],
 	images: string[],
@@ -22,33 +24,8 @@ export type Portfolio = {
 	commission: Commission,
 };
 
+export type Section = 'Android/iOS' | 'Web' | 'Illustration' | 'Photo' | 'Video';
+
 export type Commissions = {
 	[key in string]: Commissions;
-};
-
-export type Commission = {
-	client: {
-		name: string,
-		email: string,
-		phone: string,
-		nickname: string,
-		profileImage: string,
-	},
-	review: Review,
-	createdAt: Date,
-	endedAt: Date,
-	details: CommissionDetail,
-};
-
-export type Review = {
-	score: number,
-	content: string,
-};
-
-export type CommissionDetail = {
-	title: string,
-	content: string,
-	cost: number,
-	status: string,
-	deadline: Date,
 };
