+++ src/types/index.ts
@@ -10,13 +10,14 @@ export type ISectionFactory = {
 
 export type SetState<T> = React.Dispatch<React.SetStateAction<T>>;
 
-export * from '@/types/portfolio';
-export * from '@/types/user';
+// export * from '@/types/portfolio';
+// export * from '@/types/user';
 export * from '@/types/slider';
 export * from '@/types/toast';
-export * from '@/types/commission';
+// export * from '@/types/commission';
 export * from '@/types/style';
 
 export type * from '@/types/api-data/message';
-// export type * from '@/types/api-data/portfolio';
-// export type * from '@/types/api-data/user';
+export type * from '@/types/api-data/portfolio';
+export type * from '@/types/api-data/user';
+export type * from '@/types/api-data/commission';
